test: make self-hosted test available on CI
