+++ test/e2e/e2esh/main_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package e2eslow
+package e2esh
 
 import (
 	"testing"
