+++ test/e2e/main_test.go
@@ -15,25 +15,11 @@
 package e2e
 
 import (
-	"os"
 	"testing"
 
 	"github.com/coreos/etcd-operator/test/e2e/framework"
-
-	"github.com/sirupsen/logrus"
 )
 
 func TestMain(m *testing.M) {
-	if err := framework.Setup(); err != nil {
-		logrus.Errorf("fail to setup framework: %v", err)
-		os.Exit(1)
-	}
-
-	code := m.Run()
-
-	if err := framework.Teardown(); err != nil {
-		logrus.Errorf("fail to teardown framework: %v", err)
-		os.Exit(1)
-	}
-	os.Exit(code)
+	framework.MainEntry(m)
 }
